Prepare course skeleton.
Skeletons will help to create and maintain basic language courses
that can be directly shipped with Artikulate. The idea is as follows:
Artikulate will provide one course skeleton that can be used as base
for translators/teachers to create specific courses for their languages.
The so created courses are associated with the skeleton and by this
can by synced whenever the skeleton is extended.