Prepare settings for restricting use of repository.
In the future the only available resources in the editor will
come from the repository and optional a user can also use courses
from repository at trainer.

The reason is that otherwise it is very complicated to uses
downloaded courses since their recordings or phrases my change and
we must avoid loosing such changes.