Default sound build backend to QtMultimedia.
In the GStreamer 0.10 series there is a synchronization bug when
using GStreamer in combination with PulseAudio leading to deadlocks
when finishing a recording stream. This bug is exclusively fixed in
GStreamer 1.x. Though for porting to it, we require QtGStreamer to
also port to the GStreamer 1.x API.
Until this port is done, we have to default to QtMultimedia since on
most systems nowadays we find the problematic PulseAudio vs. GStreamer
deadlocks.