Remove obsolete sound plugin infrastructure and GStreamer backend
The original reason for the plugin infrastructure was due to the
GStreamer 0.10 to 1.x porting because of distros possibly shipping
incompatible versions of QtMultimedia and GStreamer. This reason does
not exist anymore. Moreover, the to expecte QtMultimedia version with
Qt6 will a future proof solution to target more operating systems than
just Linux via GStreamer.