Remove audio volume settings
Today, the expectation is to use the system volume for nearly
everything. Because of this, the volume settings were removed some time
ago, yet the automatic volume settings from existing settings was not
removed correctly, leading to situations where sound is played at
minimal volume without the user to change it.